refactor: rename types for to match the content
