+++ src/Trace/Logger.interface.ts
