+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './Constants';
 export * from './OpenTelemetryModule';
 export * from './OpenTelemetryModuleAsyncOption';
-export * from './OpenTelemetryModuleConfig';
+export * from './OpenTelemetryModuleDefaultConfig';
 
 // Trace
 export * from './Trace/Decorators/Span';
@@ -11,7 +11,7 @@ export * from './Trace/TraceWrapper';
 export * from './Trace/Injectors/ControllerInjector';
 export * from './Trace/Injectors/EventEmitterInjector';
 export * from './Trace/Injectors/GuardInjector';
-export * from './Trace/Injectors/LoggerInjector';
+export * from './Trace/Injectors/ConsoleLoggerInjector';
 export * from './Trace/Injectors/PipeInjector';
 export * from './Trace/Injectors/ScheduleInjector';
 export * from './Trace/NoopTraceExporter';
